ок

+++ metrics/cputime_nop.go
@@ -21,6 +21,6 @@ package metrics
 
 // getProcessCPUTime returns 0 on Windows as there is no system call to resolve
 // the actual process' CPU time.
-func getProcessCPUTime() int64 {
+func getProcessCPUTime() float64 {
 	return 0
 }
